Pulled db sync init into a new service. Issue: #993
